+++ foundation-models/openai/src/test/java/com/sap/ai/sdk/foundationmodels/openai/OpenAiClientTest.java
@@ -377,7 +377,9 @@ class OpenAiClientTest {
 
     verify(
         postRequestedFor(urlPathEqualTo("/embeddings"))
-            .withRequestBody(equalToJson("""
+            .withRequestBody(
+                equalToJson(
+                    """
                       {"input":["Hello World"]}""")));
   }
 
@@ -422,8 +424,7 @@ class OpenAiClientTest {
       try (Stream<OpenAiChatCompletionDelta> stream = client.streamChatCompletionDeltas(request)) {
         assertThatThrownBy(() -> stream.forEach(System.out::println))
             .isInstanceOf(OpenAiClientException.class)
-            .hasMessage(
-                "Failed to parse response from OpenAI model and error message: 'exceeded token rate limit'");
+            .hasMessage("Failed to parse response and error message: 'exceeded token rate limit'");
       }
 
       Mockito.verify(inputStream, times(1)).close();