+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/OrchestrationClient.java
@@ -5,11 +5,15 @@ import static com.sap.ai.sdk.core.JacksonConfiguration.getDefaultObjectMapper;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.annotations.Beta;
 import com.sap.ai.sdk.core.AiCoreService;
 import com.sap.ai.sdk.core.DeploymentResolutionException;
-import com.sap.ai.sdk.core.commons.ClientResponseHandler;
+import com.sap.ai.sdk.core.common.ClientResponseHandler;
+import com.sap.ai.sdk.core.common.ClientStreamingHandler;
+import com.sap.ai.sdk.core.common.StreamedDelta;
+import com.sap.ai.sdk.orchestration.model.ChatMessagesInner;
 import com.sap.ai.sdk.orchestration.model.CompletionPostRequest;
 import com.sap.ai.sdk.orchestration.model.CompletionPostResponse;
 import com.sap.ai.sdk.orchestration.model.LLMModuleResult;
@@ -47,6 +51,14 @@ public class OrchestrationClient {
     JACKSON.addMixIn(
         ModuleResultsOutputUnmaskingInner.class,
         JacksonMixins.ModuleResultsOutputUnmaskingInnerMixIn.class);
+
+    final var module =
+        new SimpleModule()
+            .addDeserializer(
+                ChatMessagesInner.class,
+                PolymorphicFallbackDeserializer.fromJsonSubTypes(ChatMessagesInner.class))
+            .setMixInAnnotation(ChatMessagesInner.class, JacksonMixins.NoneTypeInfoMixin.class);
+    JACKSON.registerModule(module);
   }
 
   @Nonnull private final Supplier<HttpDestination> destinationSupplier;
@@ -227,10 +239,10 @@ public class OrchestrationClient {
       val client = ApacheHttpClient5Accessor.getHttpClient(destination);
       val handler =
           new ClientResponseHandler<>(
-              CompletionPostResponse.class,
-              OrchestrationError.class,
-              OrchestrationClientException::new);
-      handler.JACKSON = JACKSON;
+                  CompletionPostResponse.class,
+                  OrchestrationError.class,
+                  OrchestrationClientException::new)
+              .objectMapper(JACKSON);
       return client.execute(postRequest, handler);
     } catch (DeploymentResolutionException
         | DestinationAccessException
@@ -283,8 +295,10 @@ public class OrchestrationClient {
       val destination = destinationSupplier.get();
       log.debug("Using destination {} to connect to orchestration service", destination);
       val client = ApacheHttpClient5Accessor.getHttpClient(destination);
-      return new OrchestrationStreamingHandler<>(deltaType)
-          .handleResponse(client.executeOpen(null, request, null));
+      return new ClientStreamingHandler<>(
+              deltaType, OrchestrationError.class, OrchestrationClientException::new)
+          .objectMapper(JACKSON)
+          .handleStreamingResponse(client.executeOpen(null, request, null));
     } catch (final IOException e) {
       throw new OrchestrationClientException("Request to the Orchestration service failed", e);
     }